requirements.txt
pystray==0.19.5
Pillow>=10.0.0
sounddevice>=0.4.6
numpy>=1.24.0
